backend/requirements.txt
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
gunicorn
requests
sqlparse
APScheduler
# psycopg2-binary
pandas
openpyxl
geopy
gspread